\name{param-class}
\Rdversion{1.1}
\docType{class}
\alias{param-class}
\alias{param}

\title{Class "param"}
\description{
Internal class for use with score functions.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "param" in the signature.
}
%\references{
%%  ~~put references to the literature/web site here~~
%}
\author{
John Kloke
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\linkS4class{scores}}
}
\examples{
showClass("param")
}
\keyword{classes}
